backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
